LayoutTests/fast/dom/shadow/layout-tests-can-access-shadow-expected.txt


This tests that LayoutTestController can access shadow DOM.

PASS shadow.nodeName is "#shadow-root"
PASS layoutTestController.shadowRoot(shadow) is null
PASS layoutTestController.shadowRoot(p) is null
PASS shadow.nodeName is "#shadow-root"
PASS shadow === layoutTestController.shadowRoot(p) is true
PASS layoutTestController.shadowRoot(keygen) is null
PASS successfullyParsed is true

TEST COMPLETE
